backend/requirements.txt
Django==2.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
psycopg2==2.7.7
pytz==2020.1
sqlparse==0.3.1
